fix:table이름변경 전 기능 정상작동확인
