+++ .gitignore
@@ -141,7 +141,7 @@ ENV/
 env.bak/
 venv.bak/
 
-tweet/migratoins/
+Review/migratoins/
 comment/migratoins/
 user/migratoins/
