+++ Review/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from Review.models import Review 
+
+# Register your models here.
+
+admin.site.register(Review)
