+++ Review/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+from Review import views
+
+
+urlpatterns = [
+    path('detail/', views.ReviewList.as_view(), name='review'),
+    path('detail/<int:Winedata_id>/<int:Review_id>/', views.Update.as_view(), name='update'),
+]
