+++ Winedata/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from Winedata.models import Winedata
+# Register your models here.
+
+admin.site.register(Winedata)
