+++ Winedata/tests.py
@@ -1,0 +1,3 @@
+from django.test import TestCase
+
+# Create your tests here.
