--- comment/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-from comment.models import Comment
-
-# Register your models here.
-
-admin.site.register(Comment)
