+++ templates/read.html
@@ -13,17 +13,17 @@
 
         <div id="search">   
         </div>
-        {% if tweet %}
+        {% if Review %}
         <div>
-        <p1>{{ tweet.id }}</p1>
+        <p1>{{ Review.id }}</p1>
         <div>
-        <p1>{{ tweet.name }}</p1>
+        <p1>{{ Review.name }}</p1>
         <div>
-        <p1>"{{ tweet.country }}</p1>
+        <p1>"{{ Review.country }}</p1>
         <div>
-        <p1>{{ tweet.content }}</p1>
+        <p1>{{ Review.content }}</p1>
         <div>
-        <p1><img src="{{ tweet.image }}"</p1>
+        <p1><img src="{{ Review.image }}"</p1>
             
         {% endif %}
