--- tweet/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-from tweet.models import Tweet
-# Register your models here.
-
-admin.site.register(Tweet)
