+++ user/serializer.py
@@ -90,7 +90,7 @@ class UserProfileView(serializers.ModelSerializer):
         model= User
         fields= ('username', 'email', 'profile', 'profilename')
 
-class CommentProfileView(serializers.ModelSerializer):
+class ReviewProfileView(serializers.ModelSerializer):
     class Meta:
         model= User
         fields= ("id",'profile', 'profilename')
